requirements.txt
baka==0.4.3.dev1
baka-armor==0.1.dev1
baka-tenshi==1.0.3.dev10
pyramid_mako
colander